requirements.txt
APScheduler==3.10.4
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
lxml==5.1.0
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
